feat: Completed the character setup editor window.
1. Hided all unnecessary fields and added more user-friendly fields;
2. added a weapon selection window for user-friendly sake

CharacterSetupWindow.cs:
Replaced WeaponManager.WeaponData with WeaponDataSO to hold the weapon data for the character generation session. The reason is to keep holding the weapon prefab first,then instantiate them on the generation session. Otherwise, because weapon instantiation before the generation session is on the temporary preview character instead on the final character,the final character has a "missing object" in the WeaponManager.WeaponData.weaponInstance field.